+++ src/components/kiosk/nyt/MostPopular.js
@@ -3,6 +3,9 @@ import { useState, useEffect } from 'react'
 import axios from 'axios'
 import { useAuth } from '../../../context/AuthContext'
 
+import { getDocs, collection, where, query } from 'firebase/firestore'
+import { db } from '../../../firebase'
+
 import PostCard from '../../PostCard'
 
 import data from '../../../data/nytMostPopular.json'
@@ -20,6 +23,32 @@ export default function MostPopular() {
   const [url, setUrl] = useState('')
   const [user, setUser] = useState('')
 
+  const [favorites, setFavorites] = useState([])
+
+  const favoritesCollectionRef = collection(
+    db,
+    'favorites',
+    currentUser.email,
+    currentUser.uid
+  )
+  const q = query(
+    favoritesCollectionRef,
+    where('user', '==', String(currentUser.uid))
+  )
+  /* TODO: revisit and polish this code asap */
+
+  const getFavorites = async () => {
+    const data = await getDocs(q)
+    /*       console.log(currentUser.uid) */
+    setFavorites(
+      data.docs.map((doc) => ({
+        ...doc.data(),
+        id: doc.id,
+        user: currentUser.uid,
+      }))
+    )
+  }
+
   useEffect(() => {
     /* most popular */
     axios
@@ -33,17 +62,20 @@ export default function MostPopular() {
         console.log(err)
         setMostPopulars(data)
       })
-    /* console.log(mostPopulars[0]) */
-    const info = mostPopulars[0]
 
-    /* console.log({ author, date, description, section, title, url, user }) */
+    getFavorites()
   }, [])
 
   return (
     <div className='most-populars grid-example'>
       {mostPopulars.map((post, index) => (
         <>
-          <PostCard post={post} user={currentUser} />
+          <PostCard
+            key={index}
+            post={post}
+            user={currentUser}
+            favorites={favorites}
+          />
         </>
       ))}
     </div>