+++ src/components/kiosk/nyt/TopStories/Sports.js
@@ -2,6 +2,10 @@ import { useState, useEffect } from 'react'
 import axios from 'axios'
 import PostCard from '../../../PostCard'
 import { useAuth } from '../../../../context/AuthContext'
+
+import { getDocs, collection, where, query } from 'firebase/firestore'
+import { db } from '../../../../firebase'
+
 import data from '../../../../data/nytSports.json'
 const nytTopStoriesUrl = `https://api.nytimes.com/svc/topstories/v2/sports.json?api-key=${process.env.NEXT_PUBLIC_NYT_API_KEY}`
 /* allowed values: arts, automobiles, books, business, fashion, food, health, home, insider, magazine, movies, nyregion, obituaries, opinion, politics, realestate, science, sports, sundayreview, technology, theater, t-magazine, travel, upshot, us, world */
@@ -9,6 +13,32 @@ const nytTopStoriesUrl = `https://api.nytimes.com/svc/topstories/v2/sports.json?
 export default function SportsTopStories() {
   const [posts, setPosts] = useState([])
   const { currentUser } = useAuth()
+
+  const [favorites, setFavorites] = useState([])
+
+  const favoritesCollectionRef = collection(
+    db,
+    'favorites',
+    currentUser.email,
+    currentUser.uid
+  )
+  const q = query(
+    favoritesCollectionRef,
+    where('user', '==', String(currentUser.uid))
+  )
+  /* TODO: revisit and polish this code asap */
+
+  const getFavorites = async () => {
+    const data = await getDocs(q)
+    /*       console.log(currentUser.uid) */
+    setFavorites(
+      data.docs.map((doc) => ({
+        ...doc.data(),
+        id: doc.id,
+        user: currentUser.uid,
+      }))
+    )
+  }
   useEffect(() => {
     /* top stories */
     axios
@@ -21,13 +51,14 @@ export default function SportsTopStories() {
         console.log(err)
         setPosts(data)
       })
+    getFavorites()
   }, [])
   return (
     <div>
       <div className='top-stories grid-example'>
         {posts.map((post, index) => (
           <>
-            <PostCard post={post} user={currentUser} />
+            <PostCard post={post} user={currentUser} favorites={favorites} />
           </>
         ))}
       </div>